CORE-7562 Kludge fix for switching between App grid/tile
This only needs to occur when the user has done a search for an app and switches views.
When the results come back, either the activeView will be blank or the
number of items in the listStore will be incorrect upon switching views.

Refreshing the views seems to fix this.